Updated ActivityDataStructure with more recent analysis.
I recently created a bunch of new test posts with all possible sample data variations I could think of, and did a Google Takeout export of that.
As a result, I've been able to add a couple more unique keys to the Activity Data Structure, specifically related to Collections and Polls.
Unfortunately I wasn't able to create Events anymore myself, so I haven't been able to add those yet.
If anyone is still able to create events, and is willing to create a test event with all variables filled in and all options enabled, and can share the relevant JSON data with me, please contact me either via Google+, or by opening an issue on this project.