Big restructuring/renaming of Bash scripts for exporting Blogger G+ Comments.
While moving the Bash scripts to bin/ and the functions to lib/, I also renamed the scripts to follow a naming convention.
I also removed the broken HTML generation of the Blogger G+ Comments (though I'll be adding it back in later on), and further split off the comments retrieval into more atomic scripts.

The biggest change perhaps is that I've removed the support for reading from STDIN in the scripts, as some users reported newline issues. As a result, all the scripts now rely on passing a single item as $1, so if you want to pass the results, you'll have to use command substitution or xargs.

If you have `pup` installed, a (hopefully) more reliable way of extracting ActivityIDs from Google+ Comments Widget frames will be used.