Updated the list of all possible JSON keys for Activity posts.
Of course this is only based on my own years of posts to Google+, but I'd like to think that I have made full use of the platform, with the exception of perhaps Events, so the detected keys should hopefully be as complete as possible.
If you detect a missing key, please let me know and I'll add it. If you can also provide an example of such a missing post, I'd highly appreciate it.