Added script to archive GPlus URLs with numeric ID resolving.
`bin/archive_gplus_url.sh` will check if the plus.google.com URL contains a `+CustomUsernameHandle`, and if so, resolve it to a numeric ID. It will then call `archive_url.sh` twice; once with the original URL, and once with the +CustomUserHandle replaced with the numeric UID.
Also added a test for some example URLs to test the get_user_id() function.
Finally, added some constants containing ANSI colour codes/escape sequences, and functions, for easier formatting of output.