Updated activity_data_structure.md with latest keys, as well as key from G Suite account
comments[].media.localFilePath and postKind look to have been only recently added. The former is a relative path to media file attached in a comment, the latter looks to only have EXTERNAL_SITE_COMMENT as value for JSON files in Google+ Stream/Posts, whose filename start with
'External - ', and which seem to be comments on external posts, such as Blogger posts?
postAcl.isDomainRestricted is from posts by G Suite accounts, which are restricted to their G Suite domain. Its value is "true" when the post is domain-restricted; otherwise the key is absent.