Replaced $caller_path with $PT_PATH, and source with POSIX .
This should hopefully make it easier to call scripts from other directories, as PT_PATH is always supposed to point to the Plexodus-Tools root path, and will only be set if it's not already set.
Also replaced the repeated hash tests for GNU versions of tools with ENV vars that could be overwritten if desired.